Changes to rename BaseServerPod, SonarLint suggestions and PR review comments.
