+++ common/src/main/resources/Operator.properties
@@ -326,6 +326,6 @@ WLSWH-0021=Change request to domain resource ''{0}'' causes the replica count of
   size ''{2}'' respectively
 WLSWH-0022=Scale request to cluster resource ''{0}'' cannot be honored because the replica count would exceed the \
   cluster size ''{1}''
-WLSWH-0023=Unsupported introspector environment variable(s) ''{0}'' defined in ''introspector.serverPod.env'', \
-  supported values are ''JAVA_OPTIONS'', ''USER_MEM_ARGS'', ''NODEMGR_JAVA_OPTIONS'' and ''NODEMGR_MEM_ARGS''
+WLSWH-0023=Unsupported introspector environment variable(s) {0} defined in ''introspector.serverPod.env'', \
+  supported values are {1}
