+++ json-schema-generator/src/main/java/oracle/kubernetes/json/SchemaGenerator.java
@@ -208,11 +208,7 @@ public class SchemaGenerator {
   }
 
   private boolean includeInSchema(Field field) {
-    return !isStatic(field) && !isVolatile(field) && !isDisabledFeature(field) && !ignoredField(field);
-  }
-
-  private boolean ignoredField(Field field) {
-    return field.getAnnotation(ExcludeFromSchema.class) != null;
+    return !isStatic(field) && !isVolatile(field) && !isDisabledFeature(field);
   }
 
   private boolean isStatic(Field field) {
