+++ kubernetes/crd/cluster-crd.yaml
@@ -5,7 +5,7 @@ apiVersion: apiextensions.k8s.io/v1
 kind: CustomResourceDefinition
 metadata:
   annotations:
-    weblogic.sha256: 20b61999c778a5c15c98a701b939b1b6f810647159043917af6ce8421cb6df6d
+    weblogic.sha256: 9f051b9b7805fc9100cf6490873e80b91671c7165960bfbcd6e8007ae171937f
   name: clusters.weblogic.oracle
 spec:
   group: weblogic.oracle
@@ -87,7 +87,7 @@ spec:
                 type: string
               serverPod:
                 description: Customization affecting the generation of Pods for WebLogic
-                  Server instances or the Introspector Job Pod.
+                  Server instances.
                 properties:
                   nodeName:
                     description: NodeName is a request to schedule this Pod onto a
