+++ operator/src/main/java/oracle/kubernetes/weblogic/domain/model/Introspector.java
@@ -7,7 +7,7 @@ import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 import org.apache.commons.lang3.builder.ToStringBuilder;
 
-public class Introspector extends BaseServerPodConfiguration {
+public class Introspector extends IntrospectorPodConfiguration {
 
   protected Introspector getConfiguration() {
     Introspector configuration = new Introspector();
